neonvm-runner: Move cgroup management to its own file (#1185)
We probably should remove this anyways, but in the meantime, having it
all collected in one file keeps that out of the way, and should make it
easier to remove later.
